remove Ipwatcher and move those funcs into server along with tests
