+++ controller/api/destination/server.go
@@ -1,6 +1,7 @@
 package destination
 
 import (
+	"context"
 	"encoding/json"
 	"fmt"
 	"net"
@@ -29,7 +30,6 @@ type (
 		opaquePorts   *watcher.OpaquePortsWatcher
 		profiles      *watcher.ProfileWatcher
 		trafficSplits *watcher.TrafficSplitWatcher
-		ips           *watcher.IPWatcher
 		nodes         coreinformers.NodeInformer
 
 		enableH2Upgrade     bool
@@ -82,14 +82,12 @@ func NewServer(
 	opaquePorts := watcher.NewOpaquePortsWatcher(k8sAPI, log, defaultOpaquePorts)
 	profiles := watcher.NewProfileWatcher(k8sAPI, log)
 	trafficSplits := watcher.NewTrafficSplitWatcher(k8sAPI, log)
-	ips := watcher.NewIPWatcher(k8sAPI, log)
 
 	srv := server{
 		endpoints,
 		opaquePorts,
 		profiles,
 		trafficSplits,
-		ips,
 		k8sAPI.Node(),
 		enableH2Upgrade,
 		controllerNS,
@@ -196,7 +194,7 @@ func (s *server) GetProfile(dest *pb.GetDestination, stream pb.Destination_GetPr
 
 	if ip := net.ParseIP(host); ip != nil {
 		// Get the service that the IP currently maps to.
-		svcID, err := s.ips.GetSvcID(ip.String())
+		svcID, err := getSvcID(s.k8sAPI, ip.String(), log)
 		if err != nil {
 			return err
 		}
@@ -205,7 +203,7 @@ func (s *server) GetProfile(dest *pb.GetDestination, stream pb.Destination_GetPr
 			fqn = fmt.Sprintf("%s.%s.svc.%s", service.Name, service.Namespace, s.clusterDomain)
 		} else {
 			// If the IP does not map to a service, check if it maps to a pod
-			pod, err := s.ips.GetPod(ip.String(), port)
+			pod, err := getPod(s.k8sAPI, ip.String(), port, log)
 			if err != nil {
 				return err
 			}
@@ -214,7 +212,7 @@ func (s *server) GetProfile(dest *pb.GetDestination, stream pb.Destination_GetPr
 			if pod != nil {
 				// If the IP maps to a pod, we create a single endpoint and
 				// return it in the DestinationProfile response
-				podSet := s.ips.PodToAddressSet(pod).WithPort(port)
+				podSet := podToAddressSet(s.k8sAPI, pod).WithPort(port)
 				podID := watcher.PodID{
 					Namespace: pod.Namespace,
 					Name:      pod.Name,
@@ -351,6 +349,127 @@ func (s *server) GetProfile(dest *pb.GetDestination, stream pb.Destination_GetPr
 	}
 
 	return nil
+}
+
+// getSvcID returns the service that corresponds to a Cluster IP address if one
+// exists.
+func getSvcID(k8sAPI *k8s.API, clusterIP string, log *logging.Entry) (*watcher.ServiceID, error) {
+	objs, err := k8sAPI.Svc().Informer().GetIndexer().ByIndex(watcher.PodIPIndex, clusterIP)
+	if err != nil {
+		return nil, status.Error(codes.Unknown, err.Error())
+	}
+	services := make([]*corev1.Service, 0)
+	for _, obj := range objs {
+		service := obj.(*corev1.Service)
+		services = append(services, service)
+	}
+	if len(services) > 1 {
+		conflictingServices := []string{}
+		for _, service := range services {
+			conflictingServices = append(conflictingServices, fmt.Sprintf("%s:%s", service.Namespace, service.Name))
+		}
+		log.Warnf("found conflicting %s cluster IP: %s", clusterIP, strings.Join(conflictingServices, ","))
+		return nil, status.Errorf(codes.FailedPrecondition, "found %d services with conflicting cluster IP %s", len(services), clusterIP)
+	}
+	if len(services) == 0 {
+		return nil, nil
+	}
+	service := &watcher.ServiceID{
+		Namespace: services[0].Namespace,
+		Name:      services[0].Name,
+	}
+	return service, nil
+}
+
+// getPod returns a pod that maps to the given IP address. The pod can either
+// be in the host network or the pod network. If the pod is in the host
+// network, then it must have a container port that exposes `port` as a host
+// port.
+func getPod(k8sAPI *k8s.API, podIP string, port uint32, log *logging.Entry) (*corev1.Pod, error) {
+	// First we check if the address maps to a pod in the host network.
+	addr := fmt.Sprintf("%s:%d", podIP, port)
+	hostIPPods, err := getIndexedPods(k8sAPI, watcher.HostIPIndex, addr)
+	if err != nil {
+		return nil, status.Error(codes.Unknown, err.Error())
+	}
+	if len(hostIPPods) == 1 {
+		log.Debugf("found %s:%d on the host network", podIP, port)
+		return hostIPPods[0], nil
+	}
+	if len(hostIPPods) > 1 {
+		conflictingPods := []string{}
+		for _, pod := range hostIPPods {
+			conflictingPods = append(conflictingPods, fmt.Sprintf("%s:%s", pod.Namespace, pod.Name))
+		}
+		log.Warnf("found conflicting %s:%d endpoint on the host network: %s", podIP, port, strings.Join(conflictingPods, ","))
+		return nil, status.Errorf(codes.FailedPrecondition, "found %d pods with a conflicting host network endpoint %s:%d", len(hostIPPods), podIP, port)
+	}
+
+	// The address did not map to a pod in the host network, so now we check
+	// if the IP maps to a pod IP in the pod network.
+	podIPPods, err := getIndexedPods(k8sAPI, watcher.PodIPIndex, podIP)
+	if err != nil {
+		return nil, status.Error(codes.Unknown, err.Error())
+	}
+	if len(podIPPods) == 1 {
+		log.Debugf("found %s on the pod network", podIP)
+		return podIPPods[0], nil
+	}
+	if len(podIPPods) > 1 {
+		conflictingPods := []string{}
+		for _, pod := range podIPPods {
+			conflictingPods = append(conflictingPods, fmt.Sprintf("%s:%s", pod.Namespace, pod.Name))
+		}
+		log.Warnf("found conflicting %s IP on the pod network: %s", podIP, strings.Join(conflictingPods, ","))
+		return nil, status.Errorf(codes.FailedPrecondition, "found %d pods with a conflicting pod network IP %s", len(podIPPods), podIP)
+	}
+
+	log.Debugf("no pod found for %s:%d", podIP, port)
+	return nil, nil
+}
+
+func getIndexedPods(k8sAPI *k8s.API, indexName string, podIP string) ([]*corev1.Pod, error) {
+	objs, err := k8sAPI.Pod().Informer().GetIndexer().ByIndex(indexName, podIP)
+	if err != nil {
+		return nil, fmt.Errorf("failed getting %s indexed pods: %s", indexName, err)
+	}
+	pods := make([]*corev1.Pod, 0)
+	for _, obj := range objs {
+		pod := obj.(*corev1.Pod)
+		if !podReceivingTraffic(pod) {
+			continue
+		}
+		pods = append(pods, pod)
+	}
+	return pods, nil
+}
+
+func podReceivingTraffic(pod *corev1.Pod) bool {
+	phase := pod.Status.Phase
+	podTerminated := phase == corev1.PodSucceeded || phase == corev1.PodFailed
+	podTerminating := pod.DeletionTimestamp != nil
+
+	return !podTerminating && !podTerminated
+}
+
+// podToAddressSet converts a Pod spec into a set of Addresses.
+func podToAddressSet(k8sAPI *k8s.API, pod *corev1.Pod) *watcher.AddressSet {
+	ownerKind, ownerName := k8sAPI.GetOwnerKindAndName(context.Background(), pod, true)
+	return &watcher.AddressSet{
+		Addresses: map[watcher.PodID]watcher.Address{
+			watcher.PodID{
+				Name:      pod.Name,
+				Namespace: pod.Namespace,
+			}: watcher.Address{
+				IP:        pod.Status.PodIP,
+				Port:      0, // Will be set by individual subscriptions
+				Pod:       pod,
+				OwnerName: ownerName,
+				OwnerKind: ownerKind,
+			},
+		},
+		Labels: map[string]string{"namespace": pod.Namespace},
+	}
 }
 
 ////////////
