+++ controller/api/destination/server_test.go
@@ -2,6 +2,7 @@ package destination
 
 import (
 	"fmt"
+	"strings"
 	"testing"
 
 	pb "github.com/linkerd/linkerd2-proxy-api/go/destination"
@@ -244,7 +245,6 @@ metadata:
 	opaquePorts := watcher.NewOpaquePortsWatcher(k8sAPI, log, defaultOpaquePorts)
 	profiles := watcher.NewProfileWatcher(k8sAPI, log)
 	trafficSplits := watcher.NewTrafficSplitWatcher(k8sAPI, log)
-	ips := watcher.NewIPWatcher(k8sAPI, log)
 
 	// Sync after creating watchers so that the the indexers added get updated
 	// properly
@@ -255,7 +255,6 @@ metadata:
 		opaquePorts,
 		profiles,
 		trafficSplits,
-		ips,
 		k8sAPI.Node(),
 		true,
 		"linkerd",
@@ -822,4 +821,154 @@ func toAddress(path string, port uint32) (*net.TcpAddress, error) {
 		Ip:   ip,
 		Port: port,
 	}, nil
+}
+
+func TestIpWatcherGetSvcID(t *testing.T) {
+	name := "service"
+	namespace := "test"
+	clusterIP := "10.256.0.1"
+	var port uint32 = 1234
+	k8sConfigs := fmt.Sprintf(`
+apiVersion: v1
+kind: Service
+metadata:
+  name: %s
+  namespace: %s
+spec:
+  type: ClusterIP
+  clusterIP: %s
+  ports:
+  - port: %d`, name, namespace, clusterIP, port)
+
+	t.Run("get services IDs by IP address", func(t *testing.T) {
+		k8sAPI, err := k8s.NewFakeAPI(k8sConfigs)
+		if err != nil {
+			t.Fatalf("NewFakeAPI returned an error: %s", err)
+		}
+
+		err = watcher.InitializeIndexers(k8sAPI)
+		if err != nil {
+			t.Fatalf("InitializeIndexers returned an error: %s", err)
+		}
+
+		k8sAPI.Sync(nil)
+
+		svc, err := getSvcID(k8sAPI, clusterIP, logging.WithFields(nil))
+		if err != nil {
+			t.Fatalf("Error getting service: %s", err)
+		}
+		if svc == nil {
+			t.Fatalf("Expected to find service mapped to [%s]", clusterIP)
+		}
+		if svc.Name != name {
+			t.Fatalf("Expected service name to be [%s], but got [%s]", name, svc.Name)
+		}
+		if svc.Namespace != namespace {
+			t.Fatalf("Expected service namespace to be [%s], but got [%s]", namespace, svc.Namespace)
+		}
+
+		badClusterIP := "10.256.0.2"
+		svc, err = getSvcID(k8sAPI, badClusterIP, logging.WithFields(nil))
+		if err != nil {
+			t.Fatalf("Error getting service: %s", err)
+		}
+		if svc != nil {
+			t.Fatalf("Expected not to find service mapped to [%s]", badClusterIP)
+		}
+	})
+}
+
+func TestIpWatcherGetPod(t *testing.T) {
+	podIP := "10.255.0.1"
+	hostIP := "172.0.0.1"
+	var hostPort1 uint32 = 12345
+	var hostPort2 uint32 = 12346
+	expectedPodName := "hostPortPod1"
+	k8sConfigs := []string{`
+apiVersion: v1
+kind: Pod
+metadata:
+  name: hostPortPod1
+  namespace: ns
+spec:
+  containers:
+  - image: test
+    name: hostPortContainer1
+    ports:
+    - containerPort: 12345
+      hostIP: 172.0.0.1
+      hostPort: 12345
+  - image: test
+    name: hostPortContainer2
+    ports:
+    - containerPort: 12346
+      hostIP: 172.0.0.1
+      hostPort: 12346
+status:
+  phase: Running
+  podIP: 10.255.0.1
+  hostIP: 172.0.0.1`,
+		`
+apiVersion: v1
+kind: Pod
+metadata:
+  name: pod
+  namespace: ns
+status:
+  phase: Running
+  podIP: 10.255.0.1`,
+	}
+	t.Run("get pod by host IP and host port", func(t *testing.T) {
+		k8sAPI, err := k8s.NewFakeAPI(k8sConfigs...)
+		if err != nil {
+			t.Fatalf("failed to create new fake API: %s", err)
+		}
+
+		err = watcher.InitializeIndexers(k8sAPI)
+		if err != nil {
+			t.Fatalf("initializeIndexers returned an error: %s", err)
+		}
+
+		k8sAPI.Sync(nil)
+		// Get host IP pod that is mapped to the port `hostPort1`
+		pod, err := getPod(k8sAPI, hostIP, hostPort1, logging.WithFields(nil))
+		if err != nil {
+			t.Fatalf("failed to get pod: %s", err)
+		}
+		if pod == nil {
+			t.Fatalf("failed to find pod mapped to %s:%d", hostIP, hostPort1)
+		}
+		if pod.Name != expectedPodName {
+			t.Fatalf("expected pod name to be %s, but got %s", expectedPodName, pod.Name)
+		}
+		// Get host IP pod that is mapped to the port `hostPort2`; this tests
+		// that the indexer properly adds multiple containers from a single
+		// pod.
+		pod, err = getPod(k8sAPI, hostIP, hostPort2, logging.WithFields(nil))
+		if err != nil {
+			t.Fatalf("failed to get pod: %s", err)
+		}
+		if pod == nil {
+			t.Fatalf("failed to find pod mapped to %s:%d", hostIP, hostPort2)
+		}
+		if pod.Name != expectedPodName {
+			t.Fatalf("expected pod name to be %s, but got %s", expectedPodName, pod.Name)
+		}
+		// Get host IP pod with unmapped host port
+		pod, err = getPod(k8sAPI, hostIP, 12347, logging.WithFields(nil))
+		if err != nil {
+			t.Fatalf("expected no error when getting host IP pod with unmapped host port, but got: %s", err)
+		}
+		if pod != nil {
+			t.Fatal("expected no pod to be found with unmapped host port")
+		}
+		// Get pod IP pod and expect an error
+		_, err = getPod(k8sAPI, podIP, 12346, logging.WithFields(nil))
+		if err == nil {
+			t.Fatal("expected error when getting by pod IP and unmapped host port, but got none")
+		}
+		if !strings.Contains(err.Error(), "pods with a conflicting pod network IP") {
+			t.Fatalf("expected error to be pod IP address conflict, but got: %s", err)
+		}
+	})
 }
